Move TileElementsView code into TileElementsView.h
